requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
pydantic>=2.8.0
pydantic-settings>=2.4.0
python-dateutil>=2.9.0.post0
pytz>=2024.1
Pillow>=10.0.0
google-generativeai>=0.7.2
icalendar>=5.0.12
python-dotenv>=1.0.1
